Debugged (but did not solve) issues with Audio3D
Moved init() out of main loop so that the IRFs would only be loaded
once. Also added a debug int so that we can disable debug messages if we
want. As of now, irfs load properly in init(), but calling the get_irf()
function results in an immediate segfault. Not sure why...
Also worth noting: As of now, I set it up so that it only loops twice
before stopping. We can change this back by removing the while (true) at
the bottom of runAudio3D.