Converted Convolve to AsyncTask
These changes build on Allison's first commit (before the weird merge
thing), so the code actually runs properly. The AsyncTask thing did
speed the loop up significantly, though there are still periodic blips
on the emulator (while it is waiting for garbage collection?). I'm not
sure if those blips would be audible on the PandaBoard.